Add 'my-profile' and 'current-hub' rules.
Add convenience URLs (that don’t require calculation) so that
`http://website.com/my-profile/` is redirected to
`http://website.com/members/username/`. User is routed via the login
screen if not already logged in.

Similarly, add an endpoint so that `/current-hub/action-slug` redirects
to `http://website.com/groups/group-slug/action-slug` when the request
originates in a group.